Refactor the FlinkManifest to make it more clear
